requirements.txt
numpy
Pillow
keras
flask
opencv-python
gunicorn
tensorflow

